refactor: "put all old project files into old_project folder"
